requirements.txt
pytest
pytest-playwright
python-dotenv